% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{management_population}
\alias{management_population}
\title{Title}
\usage{
management_population(NResidue = NULL, NFerlitizer = NULL,
  NPlantingDate = NULL, m = 1)
}
\arguments{
\item{m}{}
}
\description{
Title
}
